MDEV-37382 crash on CREATE OR REPLACE GTT under LOCK TABLES.
CREATE OR REPLACE GTT under LOCK TABLES was forbidden in the original
commit, as a result of rebase, however not in all the relevant places.

Crash analysis reveals that a failure is a result of a wrong TABLE
handle choice:

CREATE can open another table when CREATE...SELECT is
used, i.e. for a purpose of data reading. This is why the heuristic is
choosing to open a local table handle of Global temporary table in
CREATE TABLE.

It turns out, that CREATE OR REPLACE is another special case, when the
original table can be opened several times. Moreover, LOCK TABLES
collides with the latter as another corner case.

The solution in this patch is non-uniform: it tries to handle
unsupported cases and report error, but even that wasn't trivial.

a) CREATE OR REPLACE requires a careful decision on whether to open a
parent (global) TABLE handle, or a child (local).
It should choose a child when AS SELECT is used, for both source and
sink.
b) It also shouldn't end up in creating a new child,
if the table to replace is opened (in MDL_SHARED_UPGRADEABLE mode).

c) Additionally, LOCK TABLES + CREATE OR REPLACE end up re-opening
tables a few times with cumbersome locking rules. This becomes a big
problem, when a table was categorized as normal, and then it
re-opened as a temporary table. From open_tables, it's hard to
determine, which table should be opened in every case.

d) At the same time, CREATE OR REPLACE ends up in starting/ending a
transaction more than once, making it hard to work
with ON COMMIT DELETE ROWS (ending up with dangling pointers).

Overall in this commit:
1. Learn to determine the table to REPLACE is opened. We need a parent
table. This can be hackishly determined by MDL lock type,
see is_open_for_create_replace.
This fixes case (b).

2. Forbid CREATE OR REPLACE with ON COMMIT DELETE ROWS tables, even if
they are a part of AS SELECT.
This addresses (d)

3. Open a parent table if MYSQL_OPEN_GET_NEW_TABLE (a part of
MYSQL_OPEN_REOPEN) is in flags.

4. Forbid using global temporary tables under
LOCK TABLES + CREATE OR REPLACE + SELECT/VALUES

3 and 4 together address (c).

In all the rest cases, (a) is supported.