MDEV-37395 SIGSEGV on CREATE TABLE ... SELECT where source table is a GTT
select_create::send_eof commits a transaction, leaving a dangling
pointer to a ON COMMIT DELETE ROWS table (TABLE*).

There is a lot of access to that pointer after send_eof (also in
transaction error handling, namely abort_result_set), making it very
problematic to build a flow which would work correctly with
ON COMMIT DELETE ROWS.

So this is worked around by manually calling commit_global_tmp_tables in
the create table code in the select case.