MDEV-37541 Race of rolling back and committing transaction to binlog
The 2nd commit of two contains the fixes.

Two transactions could binlog their completions in opposite to how it
is done in Engine. That is is rare situations ROLLBACK in Engine could
be scheduled first by 2pc coordinator to leave by chance binlogging of
the transaction and whatever might be dependent on it in execution time.

This is fixed with ensuring the binlog rollback handlerton is always
executed first in 2pc.