MDEV-37541 Race of rolling back and committing transaction to binlog
The commit combines
1. a code polishing piece over a part of MDEV-21117 that
   made binlog handlerton always commit first in no 2pc cases and
2. the same rule now applies to the rollback.

The 2nd point fixes the following bug.
Two transactions could binlog their completions in opposite to how it
is done in Engine. That is is rare situations ROLLBACK in Engine could
be scheduled first by 2pc coordinator and thereby to allow for a race in
binlogging of the transaction and whatever might be dependent on it in
execution time.

An added test demonstrates how a "child" (in execution time dependency
sense) transaction could reach binlog before its parent.