MDEV-37541 Race of rolling back and committing transaction to binlog
The commit combines
1. a code polishing piece over a part of MDEV-21117 that
   made binlog handlerton always commit first in no 2pc cases and
2. the same rule now applies to the rollback.

The 2nd point fixes the following bug.
Two transactions could binlog their completions in opposite to how it
is done in Engine. That is is rare situations ROLLBACK in Engine could
be scheduled first by 2pc coordinator and thereby to allow for a race in
binlogging of the transaction and whatever might be dependent on it in
execution time.

An added test demonstrates (requires to apply
`DEBUG_SYNC(entry->thd, "before_group_commit_queue")` hunk of log.cc
to the BASE) how a "child" (in execution time dependency sense)
transaction can reach binlog before its parent.

The rollback related code of this commit differs from the commit case
in that `binlog_hton::rollback()` has to exist to satisfy its specific
`ha_rollback_to_savepoint()` invoker.