Binlog-in-engine: Fix two recovery bugs
1. The binlog recovery layer gets duplicate records in case of multi-batch
recovery, seemingly any partial mini-transaction is re-applied from the
start at a batch boundary. Thus, we need to be able to accept multiple
duplicate records belonging to the same mtr.

2. All records in a mini-transaction share the same LSN. Thus, when an mtr
spans two binlog files, the LSN of the mtr may compare larger than the
start_lsn of file_no+1, while part of that mtr (but not all of it) applies
to file_no. In this case, the record could incorrectly be interpreted as
applying to file_no+2 instead of file_no.

Both bugs found as test failures of binlog_in_engine.recover_concurrent_dml.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>